move `IntegrationCredentialsManager` to `backend.integrations.creds_store`
